backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
asyncio==3.4.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
Django==5.1.2
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-filter==24.3
django-rest-swagger==2.2.0
django-timezone-field==7.1
djangorestframework==3.15.2
drf-yasg==1.21.8
fake-useragent==1.5.1
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
kombu==5.5.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
multidict==6.1.0
openapi-codec==1.3.2
outcome==1.3.0.post0
packaging==24.1
prompt_toolkit==3.0.50
propcache==0.2.0
psycopg2==2.9.10
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
RapidFuzz==3.10.1
redis==5.2.1
requests==2.32.3
selenium==4.26.1
setuptools==75.8.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.15.2
